refactor : migration, model rename etc..

+++ database/migrations/2024_07_21_165049_create_criterias_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('criterias', function (Blueprint $table) {
             $table->id();
             $table->string('name')->comment('Nama Kriteria');
-            $table->float('weight')->comment('Bobot Kriteria');
+            $table->decimal('weight', 5, 2)->comment('Bobot Kriteria');
             $table->timestamps();
         });
     }
